feat: Add dog detector using Squeezenet
In the previous commit, we had VGG16 used to detect dog face. This network performed really well with high percentage in true positives and low in false negatives, but the network were so slow in inferencing images. In this commit, we used squeezenet to have much speed in inferencing time without sacrificing much accuracy.